# Conflicts:
#	Code/Framework/AzFramework/Platform/Mac/AzFramework/Process/ProcessWatcher_Mac.cpp
#	Code/Tools/AssetProcessor/native/utilities/BuilderManager.cpp

+++ Code/Framework/AzCore/AzCore/Math/Aabb.inl
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #pragma once
 
@@ -64,10 +60,10 @@ namespace AZ
     }
 
 
-    AZ_MATH_INLINE Aabb Aabb::CreatePoints(const Vector3* pts, int numPts)
+    AZ_MATH_INLINE Aabb Aabb::CreatePoints(const Vector3* pts, size_t numPts)
     {
         Aabb aabb = Aabb::CreateFromPoint(pts[0]);
-        for (int i = 1; i < numPts; ++i)
+        for (size_t i = 1; i < numPts; ++i)
         {
             aabb.AddPoint(pts[i]);
         }
