+++ Code/Framework/AzNetworking/AzNetworking/UdpTransport/DtlsEndpoint.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AzNetworking/UdpTransport/DtlsEndpoint.h>
 #include <AzNetworking/UdpTransport/DtlsSocket.h>
@@ -88,7 +84,7 @@ namespace AzNetworking
         if (dtlsData.GetSize() > 0)
         {
             const uint8_t* encryptedData = dtlsData.GetBuffer();
-            const uint32_t encryptedSize = dtlsData.GetSize();
+            const uint32_t encryptedSize = static_cast<uint32_t>(dtlsData.GetSize());
             BIO_write(m_readBio, encryptedData, encryptedSize);
         }
         DtlsEndpoint::HandshakeState prevState = m_state;
@@ -200,7 +196,7 @@ namespace AzNetworking
         // Need to do this... connection negotiation may have left data in the write bio that we need to send out
         if (BIO_ctrl_pending(m_writeBio) > 0)
         {
-            const uint32_t maxBufferSize = outHandshakeData.GetCapacity();
+            const uint32_t maxBufferSize = static_cast<uint32_t>(outHandshakeData.GetCapacity());
             outHandshakeData.Resize(maxBufferSize);
             const int32_t dataSize = BIO_read(m_writeBio, outHandshakeData.GetBuffer(), maxBufferSize);
             outHandshakeData.Resize(dataSize);
