+++ Code/Framework/AzNetworking/AzNetworking/UdpTransport/UdpFragmentQueue.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AzNetworking/UdpTransport/UdpFragmentQueue.h>
 #include <AzNetworking/UdpTransport/UdpConnection.h>
@@ -112,7 +108,7 @@ namespace AzNetworking
                 return true;
             }
 
-            totalPacketSize += packetFragments[index]->GetChunkBuffer().GetSize();
+            totalPacketSize += static_cast<uint32_t>(packetFragments[index]->GetChunkBuffer().GetSize());
         }
 
         // We now mark this sequence as delivered, so if by some chance all the individual chunks get redelivered again we don't double deliver the reconstructed packet
@@ -129,7 +125,7 @@ namespace AzNetworking
         uint8_t* bufferPointer = buffer.GetBuffer();
         for (uint32_t index = 0; index < packetFragments.size(); ++index)
         {
-            const uint32_t chunkSize = packetFragments[index]->GetChunkBuffer().GetSize();
+            const uint32_t chunkSize = static_cast<uint32_t>(packetFragments[index]->GetChunkBuffer().GetSize());
             memcpy(bufferPointer, packetFragments[index]->GetChunkBuffer().GetBuffer(), chunkSize);
             bufferPointer += chunkSize;
         }
@@ -137,7 +133,7 @@ namespace AzNetworking
         // We can erase all the chunks now, packet is completed
         m_packetFragments.erase(fragmentSequence);
 
-        NetworkOutputSerializer networkSerializer(buffer.GetBuffer(), buffer.GetSize());
+        NetworkOutputSerializer networkSerializer(buffer.GetBuffer(), static_cast<uint32_t>(buffer.GetSize()));
         {
             ISerializer& networkISerializer = networkSerializer; // To get the default typeinfo parameters in ISerializer
