+++ Code/Framework/AzNetworking/AzNetworking/UdpTransport/UdpReaderThread.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AzNetworking/UdpTransport/UdpReaderThread.h>
 #include <AzNetworking/UdpTransport/UdpSocket.h>
@@ -181,7 +177,7 @@ namespace AzNetworking
                 }
 
                 IpAddress address;
-                const uint32_t bufferHead = receiveBuffer.GetSize();
+                const uint32_t bufferHead = static_cast<uint32_t>(receiveBuffer.GetSize());
                 if (bufferHead + MaxUdpTransmissionUnit >= receiveBuffer.GetCapacity())
                 {
                     AZLOG_INFO("Receive buffer full, leaving data on the socket. Size exceeded by %d",
