+++ Code/Tools/AssetProcessor/native/utilities/BatchApplicationServer.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 #include <native/utilities/BatchApplicationServer.h>
 #include "native/utilities/assetUtils.h"
 
@@ -42,7 +38,7 @@ bool BatchApplicationServer::startListening(unsigned short port)
 
         // Since we're starting up builders ourselves and informing them of the port chosen, we can scan for a free port
 
-        while (!listen(QHostAddress::Any, m_serverListeningPort))
+        while (!listen(QHostAddress::Any, static_cast<quint16>(m_serverListeningPort)))
         {
             auto error = serverError();
