+++ Code/Tools/AssetProcessor/native/utilities/UnitTestShaderCompilerServer.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 #include "native/utilities/UnitTestShaderCompilerServer.h"
 #include "native/assetprocessor.h"
 #include <QTcpServer>
@@ -53,7 +49,7 @@ void UnitTestShaderCompilerServer::startServer()
 {
     if (!m_server->isListening())
     {
-        if (!m_server->listen(QHostAddress(m_serverAddress), m_serverPort))
+        if (!m_server->listen(QHostAddress(m_serverAddress), static_cast<quint16>(m_serverPort)))
         {
             AZ_TracePrintf(AssetProcessor::DebugChannel, "Server %s could not start.\n", m_serverAddress.toUtf8().data());
             emit errorMessage("Server could not start ");
