+++ Gems/AWSMetrics/Code/Tests/MetricsQueueTest.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AWSMetricsConstant.h>
 #include <AWSMetricsGemMock.h>
@@ -128,7 +124,7 @@ namespace AWSMetrics
             queue.AddMetrics(metrics);
         }
 
-        int maxCapacity = queue[0].GetSizeInBytes() * NumTestMetrics / 2;
+        int maxCapacity = static_cast<int>(queue[0].GetSizeInBytes() * NumTestMetrics / 2);
 
         ASSERT_EQ(queue.FilterMetricsByPriority(maxCapacity), NumTestMetrics / 2);
         ASSERT_EQ(queue.GetNumMetrics(), NumTestMetrics / 2);
@@ -234,7 +230,7 @@ namespace AWSMetrics
     {
         MetricsEvent metrics;
         metrics.AddAttribute(MetricsAttribute(AttrName, AttrValue));
-        int sizeOfEachMetrics = metrics.GetSizeInBytes();
+        int sizeOfEachMetrics = static_cast<int>(metrics.GetSizeInBytes());
 
         MetricsQueue queue;
         queue.AddMetrics(metrics);
