+++ Gems/Atom/Asset/Shader/Code/Source/Editor/AzslCompiler.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AzCore/Casting/numeric_cast.h>
 #include <AzCore/Component/ComponentApplicationBus.h>
@@ -23,7 +19,7 @@
 
 #include <Atom/RPI.Reflect/Shader/ShaderOptionGroupLayout.h>
 
-#include <AtomCore/Serialization/Json/JsonUtils.h>
+#include <AzCore/Serialization/Json/JsonUtils.h>
 
 #include <AzFramework/IO/LocalFileIO.h>
 #include <AzFramework/IO/FileOperations.h> // [GFX TODO] Remove when [ATOM-15472]
@@ -215,11 +211,11 @@ namespace AZ
                             case rapidjson::kNumberType:
                                 if (name == "cols")
                                 {
-                                    inputStructParams.m_variable.m_cols = itr2->value.GetInt();
+                                    inputStructParams.m_variable.m_cols = static_cast<uint8_t>(itr2->value.GetInt());
                                 }
                                 else if (name == "rows")
                                 {
-                                    inputStructParams.m_variable.m_rows = itr2->value.GetInt();
+                                    inputStructParams.m_variable.m_rows = static_cast<uint8_t>(itr2->value.GetInt());
                                 }
                                 else if (name == "semanticIndex")
                                 {
@@ -308,7 +304,7 @@ namespace AZ
                             case rapidjson::kNumberType:
                                 if (name == "cols")
                                 {
-                                    outputStructParams.m_variable.m_cols = itr2->value.GetInt();
+                                    outputStructParams.m_variable.m_cols = static_cast<uint8_t>(itr2->value.GetInt());
                                 }
                                 else if (name == "semanticIndex")
                                 {
