+++ Gems/Atom/Feature/Common/Code/Tests/SparseVectorTests.cpp
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 
 #include <AzCore/UnitTest/TestTypes.h>
 #include <Atom/Feature/Utils/SparseVector.h>
@@ -85,7 +81,7 @@ namespace UnitTest
             EXPECT_EQ(data.c, TestData::DefaultValueC);
 
             // Assign new unique values
-            data.a = TestData::DefaultValueA * i;
+            data.a = TestData::DefaultValueA * static_cast<int>(i);
             data.b = TestData::DefaultValueB * float(i);
             data.c = i % 2 == 0;
         }
@@ -194,12 +190,12 @@ namespace UnitTest
             EXPECT_EQ(data.b, TestData::DefaultValueB);
             EXPECT_EQ(data.c, TestData::DefaultValueC);
             
-            data.a = TestData::DefaultValueA * i;
+            data.a = TestData::DefaultValueA * static_cast<int>(i);
             data.b = TestData::DefaultValueB * float(i);
             data.c = i % 2 == 0;
 
             // Assign some values to the uninitialized primitive types
-            container.GetElement<1>(indices[i]) = i * 10;
+            container.GetElement<1>(indices[i]) = static_cast<int>(i * 10);
             container.GetElement<2>(indices[i]) = i * 20.0f;
         }
 
@@ -258,7 +254,7 @@ namespace UnitTest
         {
             indices[i] = container.Reserve();
 
-            container.GetElement<1>(i) = i * 10;
+            container.GetElement<1>(i) = static_cast<int>(i * 10);
             container.GetElement<2>(i) = i * 20.0f;
         }
