+++ Gems/Atom/RHI/Code/Include/Atom/RHI/ImageProperty.h
@@ -1,14 +1,10 @@
 /*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
 #pragma once
 
 #include <Atom/RHI.Reflect/ImageDescriptor.h>
@@ -187,7 +183,7 @@ namespace AZ
                     else
                     {
                         // Insert intervals by mip level.
-                        for (uint32_t mipLevel = subResourceRange.m_mipSliceMin; mipLevel <= subResourceRange.m_mipSliceMax; ++mipLevel)
+                        for (uint16_t mipLevel = subResourceRange.m_mipSliceMin; mipLevel <= subResourceRange.m_mipSliceMax; ++mipLevel)
                         {
                             m_intervalMap.assign(
                                 ConvertSubresourceToIndex(aspect, mipLevel, subResourceRange.m_arraySliceMin),
@@ -277,7 +273,7 @@ namespace AZ
                     else
                     {
                         // Traverse one mip level at a time.
-                        for (uint32_t mipLevel = subResourceRange.m_mipSliceMin; mipLevel <= subResourceRange.m_mipSliceMax; ++mipLevel)
+                        for (uint16_t mipLevel = subResourceRange.m_mipSliceMin; mipLevel <= subResourceRange.m_mipSliceMax; ++mipLevel)
                         {
                             getIntervals(
                                 ConvertSubresourceToIndex(aspect, mipLevel, subResourceRange.m_arraySliceMin),
@@ -336,8 +332,8 @@ namespace AZ
         {
             const uint32_t subresourcesPerAspect = m_imageDescriptor.m_mipLevels * m_imageDescriptor.m_arraySize;
             return ImageSubresource(
-                (index % subresourcesPerAspect) / m_imageDescriptor.m_arraySize,
-                (index % subresourcesPerAspect) % m_imageDescriptor.m_arraySize,
+                static_cast<uint16_t>((index % subresourcesPerAspect) / m_imageDescriptor.m_arraySize),
+                static_cast<uint16_t>((index % subresourcesPerAspect) % m_imageDescriptor.m_arraySize),
                 static_cast<ImageAspect>(index/ subresourcesPerAspect));
         }
